requirements.txt
numpy
opencv-python
dxcam
ultralytics
easyocr
pygetwindow
pyautogui
Pillow
